Split adbd.te into adbd.te and adbd_common.te.
In trade-in mode, adbd will setcon() itself to a much lower privileged
sepolicy so that only a single command will work: "adb shell
tradeinmode". To make this work we are refactoring adbd.te.

This splits the most critical parts of adbd.te into a separate file,
adbd_common.te, so that it is clear which permissions need to be
inherited by both normal adbd and trade-in mode.

adbd_common.te was created by running adbd in a context with no
privileges, and then iteratively moving lines from adbd.te based on
audit messages, until "adb shell" could drop into a totally useless
shell.

Ideally, this change would have been to introduce a new attribute that
could be inherited. However trunk stable makes this exceedingly
difficult, so instead we opt for a small amount of duplication.

Bug: 307713521
Test: adb still works on cuttlefish
Change-Id: Ia77f9ea3a12e697805b331c1379421d124e013d2