aconfigd: cleanup
Various cleanups of aconfigd

1) Remove unnecessary aconfigd permissions for /dev/kmsg.

This file descriptor is opened by init and passed to the process,
so permissions like "open" are not needed. The only permission
needed is "write".

2) Get rid of aconfigd's ability to connect to its own
/dev/socket/aconfigd socket.

A process connecting to itself is unlikely to be useful and most
likely indicates a bug. I'm struggling to imagine situations where
such recursive connections would be useful. This appears to be an
artifact of early aconfigd bringup and unused by current code.

3) Move neverallow assertions

By convention, SELinux neverallow assertions go at the end of the
relevant *.te file.

4) Move typeattribute to type line

This makes the policy simpler to read.

5) Remove shell access to the aconfigd socket.

This was already surrounded by a userdebug_or_eng block. Remove
it entirely from policy to avoid confusion and reduce build-time
variation of SELinux policy rules.

Bug: 370499640
Test: Build compiles and boots with no SELinux errors
Test: adb root && \
      m aconfigd_socket_test && \
      adb sync data && \
      adb shell /data/nativetest64/aconfigd_socket_test/aconfigd_socket_test
Change-Id: I29d9d160cbd7223771cb8ae17d9e477d5d5b8ac6