domain.te: tmpfs neverallows
By default, tmpfs filesystems mounted in Android are labeled with
the tmpfs security label. Code which mounts these filesystems generally
assigns a more specific security label for use by domains.

Prohibit the use of the generic tmpfs security label when creating
files. Instead, encourage the use of more fine grain security labels.

Add a few exceptions where policy currently allows the creation
of some types of file objects. It's not clear to me if these
exceptions are still needed, but they are currently allowed by
policy.

Not covered is the use of symbolic links on tmpfs filesystems,
which still appear to be in common use. This is especially true
for the /mnt/sdcard symlink which has been deprecated since the
Gingerbread days.

Test: compiles (neverallow assertions are compile only checks)
Change-Id: I4bc9d186c53a5dc435813999f69539bcabe2d942