Mark selinux_policy_system_ext as system_ext_specific
As part of the make to soong conversion, we are currently autogenerating
android_filesystem soong modules for each partition (system, system_ext,
vendor, ...), with the goal to eventually substitute the kati built
partitions. In order to generate the deps of these partitions, we are
currently using a heuristic to classify the modules in
`PRODUCT_PACKAGES` with the associated partition.

Mark the phony module `selinux_policy_system_ext` as
`system_ext_specific` to help with this conversion heursitic.

Test: Verified that the required modules of `selinux_policy_system_ext`
are installed in the autogenerated
`aosp_cf_x86_64_phone_generated_system_ext_image` system_ext partition
module.

Bug: 372487849
Change-Id: I436ece3619cc0a023830f7b353e89a25f3d54a27