Refactor selinux_policy_nonsystem to separate phonies for vendor/odm
As part of the mk->bp migration, we are currently using heuristics to
classify the packages in `PRODUCT_PACKAGES` to deps of the different
android_filesystem modules. To aid such heuristics, separate phony
targets will be created for selinux_policy_vendor and
selinux_policy_odm. The former wil be added to deps of the autogenerated
vendor module and the latter will be added to deps of the autogenerated
odm module

Test: verified that the soong build vendor.img and odm.img have the same
files as the kati vendor.img and odm.img for aosp_cf_x86_64_phone

Bug: 376755916
Change-Id: Icba06fd978e44d7c8d293ba3d85853df04f907dd