[Thread] allow ot-daemon to read/write system_server created UDP socket
This is similar to the rawip_socket but it's UDP socket here. The use
case is to allow the ot-daemon being the DNS server for the Thread
network which needs to bind to port 53. The ot-daemon service won't bind
the port itself as it's an unprivileged process, instead, the
system_server will create the socket for it.

Bug: 379998097
Test: Tested with cuttlefish
Change-Id: I414a36696e5a219ad932289f5eade02652b4ecc6