Allow init to write to /sys/module/dm_bufio/parameters/max_age_seconds
Label max_age_seconds with sysfs_dm_verity so that init can write to it
to set the dm-verity cache timeout.

Technically this file is for dm-bufio, not for dm-verity specifically;
writing to it affects any device-mapper target that uses dm-bufio.  But
since dm-verity is the only user of dm-bufio on Android devices, it
doesn't seem worth adding a new label.  Let's just consider it to be
another dm-verity tunable similar to prefetch_cluster.

Bug: 335233956
Test: cat /sys/module/dm_bufio/parameters/max_age_seconds
Change-Id: If8901e7f5aa4e8e3dd81639792cd4a4cd1273bfa