introducing unconstrained_vsock_violators
Adding attribute separately from uses of it so
that it can be applied in different places.

Basically, SELinux doesn't have a good view on
how vsock connections are setup, and they are
unconstrained. We need to limit these and either
allow SELinux to understand what's on the other
side of the connection, or delegate the permission
model to virtualizationmanager.

Bug: 347661724
Test: N/A
Merged-In: Ie0ede16fe73f609386275ed18f4b2ffe49620b12
Change-Id: Ie0ede16fe73f609386275ed18f4b2ffe49620b12