Enforce unconstrained_vsock_violators
Basically, SELinux doesn't have a good view on
how vsock connections are setup, and they are
unconstrained. We need to limit these and either
allow SELinux to understand what's on the other
side of the connection, or delegate the permission
model to virtualizationmanager.

This change puts this check in place. Additional
CLs will add tests.

Bug: 347661724
Test: build (neverallow enforced at build)
Change-Id: If59b5d51cfead32895a88cb6b0c9e39743f98ea6