Give apps read permissions on its own profile directory.
The ART profiles are saved by the ART runtime on behalf of the app,
using atomic file replacement (the profile is saved as a temporary
file and renamed to the final location).

After the rename, we want to flush the directory, so that the changes
are reflected immediately. The flushing requires read permissions on the
directory.

This is particularly useful on a force save, where app developers may
want to dump the saved profile through adb immediately.

Bug: 398982550
Test: Presubmit
Change-Id: I404414afe7a6f1e3295903c71ec3f3b7f8c74056